+++ CHANGELOG.md
@@ -5,3 +5,25 @@
 
 - Publish to jitpack
 - Initial version
+
+### Version 1.1.0
+#### Date: September 13, 2021
+
+- New Activity Extensions Added
+    setStatusBarColor
+    setNavigationBarColor
+    setSystemBarColor
+
+- New Misc Extension Added
+    isColorDark
+
+- New Graphics Extension Added
+    addGradient
+
+- New String Extensions Added
+    haveAnyAlphabetNumber
+    isContainNumber
+
+- New Validation Extensions Added
+    isPasswordValid
+    isValidPackageName